requirements.txt
requests
descope
python-dotenv